requirements.txt
psycopg2-binary
requests
beautifulsoup4
streamlit

